pyproject.toml
[project]
name = "eleventwilio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "psycopg2>=2.9.10",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.26",
    "twilio>=9.8.3",
    "uvicorn>=0.37.0",
    "websockets>=15.0.1",
]
